upate paramers types

+++ src/memory_gpu.cuh
@@ -1,3 +1,3 @@
 #pragma once
 	__global__
-	void memory_set_kernel(iris_real* ptr, size_t n, iris_real val);
+	void memory_set_kernel(iris_real* ptr, int n, iris_real val);
